seperate tensorizer related functions to independent file & nits fix
